backend/TARA_Platform/requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.14,<3.17
djangorestframework-simplejwt>=5.5,<6.0
django-cors-headers>=4.0,<5.0
drf-spectacular>=0.29,<1.0

channels>=4.0,<5.0
daphne>=4.0,<5.0
channels_redis>=4.0,<5.0
asgiref>=3.6,<4.0

psycopg2-binary>=2.9,<3.0
dj-database-url>=2.0,<4.0

requests>=2.31,<3.0
httpx>=0.25,<0.29
aiohttp>=3.9,<4.0
websockets>=12.0,<16.0

cryptography>=41.0,<47.0
PyJWT>=2.8,<3.0

python-dotenv>=1.0,<2.0
pytz>=2023.3
sqlparse>=0.5,<0.6

openai>=1.0,<3.0
fastmcp>=0.1,<3.0

uvicorn>=0.24,<0.39
streamlit>=1.28,<2.0

psutil